+++ plugin-niassa/src/main/java/ca/on/oicr/gsi/shesmu/niassa/LockLogger.java
@@ -1,7 +1,0 @@
-package ca.on.oicr.gsi.shesmu.niassa;
-
-import ca.on.oicr.gsi.provenance.model.LimsKey;
-
-public interface LockLogger {
-  void log(LimsKey key, String message);
-}